+++ CHANGES.md
@@ -1,3 +1,11 @@
+#### 0.8.1 / 2018-05-11
+* added `subheading` attribute
+* added new fields types: `number`, `tel`, `url`
+* improved  js error processing
+* added support of `carbon-copy` attribute
+* updated language dictionaries
+* css optimization
+
 #### 0.7.6 / 2018-03-20
 * added language dictionaries for `en` and `ru`
 * bug fixing